+++ example/lib/example5_themes_playground/theme/flex_theme_dark.dart
@@ -4,6 +4,7 @@ import 'package:flutter/material.dart';
 import '../../shared/const/app_color.dart';
 import '../../shared/const/app_data.dart';
 import '../../shared/controllers/theme_controller.dart';
+import 'code_theme.dart';
 
 /// This function calls [flexColorSchemeDark] and uses
 /// [FlexColorScheme.toTheme] to return the [ThemeData] object represented by
@@ -19,11 +20,24 @@ import '../../shared/controllers/theme_controller.dart';
 /// using the exact same [ColorScheme], but using just default [ThemeData] with
 /// no [FlexColorScheme] theming applied.
 ///
-ThemeData flexThemeDark(ThemeController controller) =>
-    flexColorSchemeDark(controller).toTheme;
+ThemeData flexThemeDark(ThemeController controller) {
+  // Must get the effective primary color, so we can use it as source color to
+  // harmonize the CodeTheme extension colors towards it, by using
+  // MaterialColorUtilities function Blend.harmonize, at this point
+  // the source color does not matter.
+  final Color source =
+      flexColorSchemeDark(controller, Colors.black).toScheme.primary;
+  // Now we can use our function that takes theme controller and source color,
+  // which is the effective primary color.
+  return flexColorSchemeDark(controller, source).toTheme;
+}
 
 /// Create the FlexColorScheme object represented by our current
-/// [ThemeController] configuration.
+/// [ThemeController] value [controller] configuration and [source] color.
+///
+/// The [source] color is used to color adjust all the custom code highlight
+/// theme colors added via ThemeData extension [CodeTheme], towards the
+/// [source] color.
 ///
 /// This setup may seem complex, but all the controller does is represent
 /// configuration values selected in the UI that are input to a large number
@@ -37,7 +51,7 @@ ThemeData flexThemeDark(ThemeController controller) =>
 /// The function only contains tutorial comments on topics not covered in the
 /// almost identical `flexColorSchemeLight` function. Please see it for
 /// more tutorial comments.
-FlexColorScheme flexColorSchemeDark(ThemeController controller) {
+FlexColorScheme flexColorSchemeDark(ThemeController controller, Color source) {
   final bool useBuiltIn = controller.schemeIndex > 2 &&
       controller.schemeIndex < AppColor.schemes.length - 1;
   final int flexScheme = controller.schemeIndex - 3;
@@ -226,5 +240,9 @@ FlexColorScheme flexColorSchemeDark(ThemeController controller) {
     // TODO(rydmike): Remove when fix for issue #10386 has landed in stable.
     textTheme: fakeM2TypographyTextTheme,
     primaryTextTheme: fakeM2TypographyTextTheme,
+    // Add a custom theme extension with light mode code highlight colors.
+    extensions: <ThemeExtension<dynamic>>{
+      CodeTheme.dynamic(source, Brightness.dark),
+    },
   );
 }